Move distance calculator to a common package
