+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/core/CoreAliases.kt
@@ -16,8 +16,6 @@ typealias CoreSuggestAction = com.mapbox.search.internal.bindgen.SuggestAction
 typealias CoreSearchAddress = com.mapbox.search.internal.bindgen.SearchAddress
 typealias CoreReverseMode = com.mapbox.search.internal.bindgen.ReverseMode
 typealias CoreReverseGeoOptions = com.mapbox.search.internal.bindgen.ReverseGeoOptions
-typealias CoreDistanceCalculator = com.mapbox.search.internal.bindgen.DistanceCalculator
-typealias CoreDistanceCalculatorInterface = com.mapbox.search.internal.bindgen.DistanceCalculatorInterface
 typealias CoreUserRecordsLayer = com.mapbox.search.internal.bindgen.UserRecordsLayer
 typealias CoreUserRecord = com.mapbox.search.internal.bindgen.UserRecord
 typealias CoreApiType = com.mapbox.search.internal.bindgen.ApiType
