+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/utils/extension/Point.kt
@@ -1,0 +1,14 @@
+package com.mapbox.search.base.utils.extension
+
+import com.mapbox.geojson.Point
+import com.mapbox.search.common.DistanceCalculator
+
+/**
+ * Calculates distance to specified destination point.
+ *
+ * @return distance to [destination] in meters.
+ */
+@JvmSynthetic
+fun Point.distanceTo(destination: Point): Double =
+    DistanceCalculator.instance(latitude = latitude())
+        .distance(this, destination)
